requirements.txt
fastapi[all]
uvicorn[standard]
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
pymysql
pytest
requests
websockets
firebase-admin
gunicorn
psycopg2-binary
fastapi-utils
alembic